+++ CHANGES.rst
@@ -4,7 +4,17 @@ CHANGES
 0.9.12 (unreleased)
 -------------------
 
-- Nothing changed yet.
+  API Changes
+  ~~~~~~~~~~~
+
+  - astropy.wcs.WCS no longer contains information about the original image
+    size.  Any attempt to instantiate a FITSFigure from a WCS object will raise
+    a DeprecationException.  A workaround is to add `naxisn` attributes to your
+    WCS object::
+
+       mywcs = wcs.WCS(header)
+       mywcs.naxis1 = header['NAXIS1']
+       mywcs.naxis2 = header['NAXIS2']
 
 
 0.9.11 (2013-11-29)